Support build-time key remapping
As the removed TODO hints, being able to press some keys which are not on a
standard keyboard can be useful, as is the case with KEY_POWER on Android, which
will turn on/turn off the display.

To be able to specify the key mapping using the identifiers (macros) defined in
linux/input.h, the C preprocessor is used to expand the macros to a format which
can then be interpreted by awk to generate a lookup function.