forward_input: Wait for ungrab before attempting grab
When invoking forward_input from a tool such as xbindkeys, the keyboard is
already grabbed, which naturally causes forward_input to fail in locking the
keyboard and thus receiving any XKeyEvents.

If no such events should be posted, the program is still easy to kill, since no
grab was successful.