Properly generate dependencies on generated sources
By utilizing -MG (which is incompatible with -MD) gcc can generate dependency
files for sources even though all header files are not present. This comes at
the cost of having to call gcc twice (once for dependency generation) whenever
sources change. As the compile time is so short, this is not an issue.